refactor: abstracted truncateTitle function as a reusable function
