+++ apps/web/src/helpers/truncate-title.ts
@@ -1,0 +1,10 @@
+export const truncateTitle = (title: string, maxLength: number) => {
+  if (title.length <= maxLength) {
+    return title;
+  }
+
+  const start = title.slice(0, maxLength / 2);
+  const end = title.slice(-maxLength / 2);
+
+  return `${start}.....${end}`;
+};
